Fix mypy errors in api, tests, trading handler; use real PyJWT types
- websocket_manager: remove two dead method definitions silently shadowed
  by later same-name definitions (the later ones are what runs and what
  tests exercise)
- repo tests: replace removed UsersRepo.ensure_user_for_telegram /
  update_telegram_profile with a conftest helper built on repo primitives
  (the convenience methods moved to UsersService)
- services conftest: add missing kestrel repo to ReposBundle construction
- auth: install PyJWT (inline types; encode returns str) and drop the
  v1-era types-PyJWT stub that claimed bytes
- screener_config_loader: empty YAML loads as None -> default to {}
- strategy_handler/file_data_feed/unified_analytics: None guards and
  float coercions

+++ src/data/feed/file_data_feed.py
@@ -432,8 +432,8 @@ class FileDataFeed(bt.feed.DataBase):
             # Generate realistic OHLC from close price
             volatility = abs(np.random.normal(0, 0.01))  # Intrabar volatility
 
-            high = close_price * (1 + volatility)
-            low = close_price * (1 - volatility)
+            high = float(close_price * (1 + volatility))
+            low = float(close_price * (1 - volatility))
 
             # Open is previous close (with some gap)
             if i == 0:
@@ -443,8 +443,8 @@ class FileDataFeed(bt.feed.DataBase):
                 open_price = prices[i - 1] * (1 + gap)
 
             # Ensure OHLC relationships are valid
-            high = max(high, open_price, close_price)
-            low = min(low, open_price, close_price)
+            high = max(high, float(open_price), float(close_price))
+            low = min(low, float(open_price), float(close_price))
 
             # Generate volume
             volume = np.random.randint(1000, 10000)
